requirements.txt
streamlit
python-dotenv
langchain
langchain-core
langchain-groq
langchain-huggingface
langchain-community
faiss-cpu
sentence-transformers
torch